Hardcode - Modal, C&P Dropdowns, C&P Collapse
I Tried to write as much as I can about the code and
try to make soemthing new and useful for Modal page.

To display the different types of Modal I made a way to display all the
different sizes of modal available.

When making the Dropdowns and Collapse Page I just
copy pasted the content from w3schools which I should not
have done in the first place because I do not understand the
workings of the Dropdowns and the Collapse.